573 Remove redundant primary bootloader Rename secondary one to bootloader
