--- bootloader-primary/src/main/java/ru/videmanmc/launcher/bootloader/primary/BootloaderPrimary.java
@@ -1,9 +1,0 @@
-package ru.videmanmc.launcher.bootloader.primary;
-
-public class BootloaderPrimary {
-
-    public static void main(String[] args) {
-        System.out.println("Hello World!");
-    }
-
-}
