--- bootloader-secondary/build.gradle
@@ -1,36 +1,0 @@
-plugins {
-    id 'application'
-    id 'launcher.java-build'
-    id 'launcher.version-management'
-}
-
-application {
-    mainClass = "ru.videmanmc.launcher.bootloader.secondary.BootloaderSecondary"
-}
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation project(':constants')
-    implementation project(':http-client')
-
-    compileOnly "org.jetbrains:annotations:$annotationsVer"
-
-    compileOnly "org.projectlombok:lombok:$lombokVer"
-    annotationProcessor "org.projectlombok:lombok:$lombokVer"
-
-    testImplementation(platform("org.junit:junit-bom:5.11.3"))
-    testImplementation 'org.junit.jupiter:junit-jupiter'
-    testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
-
-    testImplementation "org.mockito:mockito-core:$mockitoVer"
-    testImplementation "org.mockito:mockito-junit-jupiter:$mockitoVer"
-
-    testImplementation "com.google.jimfs:jimfs:$jimfsVer"
-}
-
-tasks.withType(Test).configureEach {
-    useJUnitPlatform()
-}
