+++ bootloader/build.gradle
@@ -1,0 +1,36 @@
+plugins {
+    id 'application'
+    id 'launcher.java-build'
+    id 'launcher.version-management'
+}
+
+application {
+    mainClass = "ru.videmanmc.launcher.bootloader.Bootloader"
+}
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    implementation project(':constants')
+    implementation project(':http-client')
+
+    compileOnly "org.jetbrains:annotations:$annotationsVer"
+
+    compileOnly "org.projectlombok:lombok:$lombokVer"
+    annotationProcessor "org.projectlombok:lombok:$lombokVer"
+
+    testImplementation(platform("org.junit:junit-bom:5.11.3"))
+    testImplementation 'org.junit.jupiter:junit-jupiter'
+    testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
+
+    testImplementation "org.mockito:mockito-core:$mockitoVer"
+    testImplementation "org.mockito:mockito-junit-jupiter:$mockitoVer"
+
+    testImplementation "com.google.jimfs:jimfs:$jimfsVer"
+}
+
+tasks.withType(Test).configureEach {
+    useJUnitPlatform()
+}
