+++ settings.gradle
@@ -4,5 +4,4 @@ include('core',
         'gui',
         'constants',
         'http-client',
-        'bootloader-secondary',
-        'bootloader-primary')
+        'bootloader')
